Rename Thompson_PatrickLab1Final.cpp to Swim School Schedule Maker and FIle Saver
This file gives two swim instructors named Jeff and Anna a schedule. The application is given a default go to schedule from the get-go, but the app lets the user change the schedule and/or save it to a .txt file. If need be, the user can reload an old schedule from the same file into a new instance of the program. This app also lets the user see if there are any open slots during the week, and also if there are any slots during the week where both instructors are available for a duo class.